+++ packages/portal-proto/package.json
@@ -40,7 +40,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-icons": "^4.12.0",
-    "react-markdown": "^8.0.3",
+    "react-markdown": "^9.0.1",
     "react-modal": "^3.14.3",
     "react-plotly.js": "^2.6.0",
     "react-redux": "^7.2.3",
